+++ ydb/public/lib/ydb_cli/commands/topic_workload/topic_workload_keyed_writer.cpp
@@ -126,6 +126,9 @@ std::shared_ptr<TTopicWorkloadKeyedWriterProducer> TTopicWorkloadKeyedWriterWork
 
     NYdb::NTopic::TProducerSettings settings;
     settings.Codec((NYdb::NTopic::ECodec)Params.Codec);
+    if (Params.BatchInnerCodec.Defined()) {
+        settings.BatchInnerCodec((NYdb::NTopic::ECodec)*Params.BatchInnerCodec);
+    }
     settings.Path(Params.TopicName);
     settings.ProducerIdPrefix(producerId);
     settings.MaxBlockTimeout(TDuration::Max());